requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
python-dotenv==1.2.1
bcrypt==5.0.0
pymongo==4.15.3
dnspython==2.4.2
gunicorn==21.2.0
requests>=2.28.0

